refactor film-card reviews component
